feat: 数据库后端支持 PostgreSQL
- 新增 --database-url / BILI_SYNC_DATABASE_URL 连接串参数，支持
postgres:// 后端；
  SQLite 行为保持兼容，created_at 一律按 UTC
存储、日期筛选/分桶按本地时区
- 实体与迁移层适配 PostgreSQL（列类型、COALESCE 唯一索引、UTC 默认值）
- 状态位域统一 i64 载体；视频源规则支持显式清空并保留既有规则
- 新增 scripts/db-migrate/ SQLite→PostgreSQL 迁移实测工具链（pgloader
驱动，
  表清单与 cast 规则全动态推导，行数对账/序列对齐/DDL 对比）
- 工具链凭证安全：密码经 PGPASSWORD 传递、临时目录 0700 自动清理、.load
0600
- 测试：SQLite 兼容回归（created_at UTC 语义、dashboard SQLite 分支）

+++ crates/bili_sync/src/api/helper.rs
@@ -35,33 +35,33 @@ impl ValidationFilter {
 }
 
 pub trait VideoRecord {
-    fn as_id_status_tuple(&self) -> (i32, u32);
+    fn as_id_status_tuple(&self) -> (i32, i64);
 }
 
 pub trait PageRecord {
-    fn as_id_status_tuple(&self) -> (i32, u32);
+    fn as_id_status_tuple(&self) -> (i32, i64);
 }
 
 impl VideoRecord for VideoInfo {
-    fn as_id_status_tuple(&self) -> (i32, u32) {
+    fn as_id_status_tuple(&self) -> (i32, i64) {
         (self.id, self.download_status)
     }
 }
 
 impl VideoRecord for SimpleVideoInfo {
-    fn as_id_status_tuple(&self) -> (i32, u32) {
+    fn as_id_status_tuple(&self) -> (i32, i64) {
         (self.id, self.download_status)
     }
 }
 
 impl PageRecord for PageInfo {
-    fn as_id_status_tuple(&self) -> (i32, u32) {
+    fn as_id_status_tuple(&self) -> (i32, i64) {
         (self.id, self.download_status)
     }
 }
 
 impl PageRecord for SimplePageInfo {
-    fn as_id_status_tuple(&self) -> (i32, u32) {
+    fn as_id_status_tuple(&self) -> (i32, i64) {
         (self.id, self.download_status)
     }
 }
@@ -110,7 +110,7 @@ where
 
 async fn execute_video_update_batch(
     txn: &DatabaseTransaction,
-    videos: impl Iterator<Item = (i32, u32)>,
+    videos: impl Iterator<Item = (i32, i64)>,
 ) -> Result<(), sea_orm::DbErr> {
     let values = videos.map(|v| format!("({}, {})", v.0, v.1)).join(", ");
     if values.is_empty() {
@@ -130,7 +130,7 @@ async fn execute_video_update_batch(
 
 async fn execute_page_update_batch(
     txn: &DatabaseTransaction,
-    pages: impl Iterator<Item = (i32, u32)>,
+    pages: impl Iterator<Item = (i32, i64)>,
 ) -> Result<(), sea_orm::DbErr> {
     let values = pages.map(|p| format!("({}, {})", p.0, p.1)).join(", ");
     if values.is_empty() {
